move execution receipt type to its own module
