+++ crates/pallet-domains/src/tests.rs
@@ -29,11 +29,12 @@ use sp_domains::bundle::bundle_v1::BundleV1;
 use sp_domains::bundle::{
     BundleHeader, BundleVersion, InboxedBundle, OpaqueBundle, SealedBundleHeader,
 };
+use sp_domains::execution_receipt::ExecutionReceipt;
 use sp_domains::merkle_tree::MerkleTree;
 use sp_domains::storage::RawGenesis;
 use sp_domains::{
-    ChainId, DomainId, ExecutionReceipt, OperatorAllowList, OperatorId, OperatorPair,
-    ProofOfElection, RuntimeId, RuntimeType,
+    ChainId, DomainId, OperatorAllowList, OperatorId, OperatorPair, ProofOfElection, RuntimeId,
+    RuntimeType,
 };
 use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_runtime::generic::{EXTRINSIC_FORMAT_VERSION, Preamble};
