+++ crates/sc-domains/src/domain_block_er/execution_receipt_protocol.rs
@@ -23,7 +23,8 @@ use sc_network::request_responses::{IncomingRequest, OutgoingResponse};
 use sc_network::{NetworkBackend, PeerId};
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
-use sp_domains::{DomainId, DomainsApi, ExecutionReceiptFor};
+use sp_domains::execution_receipt::ExecutionReceiptFor;
+use sp_domains::{DomainId, DomainsApi};
 use sp_runtime::traits::{Block as BlockT, Header};
 use std::marker::PhantomData;
 use std::sync::Arc;
