+++ crates/sc-domains/src/domain_block_er/receipt_receiver.rs
@@ -13,7 +13,8 @@ use sc_network::{IfDisconnected, NetworkRequest, PeerId, RequestFailure};
 use sc_network_sync::SyncingService;
 use sp_blockchain::HeaderBackend;
 use sp_core::{Hasher, KeccakHasher};
-use sp_domains::{DomainId, ExecutionReceiptFor};
+use sp_domains::DomainId;
+use sp_domains::execution_receipt::ExecutionReceiptFor;
 use sp_runtime::traits::{Block as BlockT, Header};
 use std::collections::BTreeMap;
 use std::marker::PhantomData;
