+++ crates/sp-domains-fraud-proof/benches/fraud_proof_verification.rs
@@ -12,8 +12,9 @@ use sc_service::{BasePath, Role};
 use sp_api::ProvideRuntimeApi;
 use sp_core::{H256, Pair as _};
 use sp_domains::bundle::{BundleValidity, InvalidBundleType};
+use sp_domains::execution_receipt::{ExecutionReceiptFor, Transfers};
 use sp_domains::merkle_tree::MerkleTree;
-use sp_domains::{ChainId, DomainsApi, ExecutionReceiptFor, Transfers};
+use sp_domains::{ChainId, DomainsApi};
 use sp_domains_fraud_proof::fraud_proof::DomainRuntimeCodeAt;
 use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::{FraudProofV1, FraudProofVariantV1};
 use sp_domains_fraud_proof::storage_proof::{BasicStorageProof, DomainRuntimeCodeProof};
