+++ crates/subspace-fake-runtime-api/src/lib.rs
@@ -10,9 +10,10 @@ use sp_core::crypto::KeyTypeId;
 use sp_core::{H256, OpaqueMetadata};
 use sp_domains::bundle::BundleVersion;
 use sp_domains::bundle_producer_election::BundleProducerElectionParams;
+use sp_domains::execution_receipt::ExecutionReceiptFor;
 use sp_domains::{
-    DomainAllowlistUpdates, DomainId, DomainInstanceData, ExecutionReceiptFor, OperatorId,
-    OperatorPublicKey, PermissionedActionAllowedBy,
+    DomainAllowlistUpdates, DomainId, DomainInstanceData, OperatorId, OperatorPublicKey,
+    PermissionedActionAllowedBy,
 };
 use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_domains_fraud_proof::storage_proof::FraudProofStorageKeyRequest;
@@ -176,7 +177,7 @@ sp_api::impl_runtime_apis! {
         }
 
         fn submit_receipt_unsigned(
-            _singleton_receipt: sp_domains::SealedSingletonReceipt<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
+            _singleton_receipt: sp_domains::execution_receipt::SealedSingletonReceipt<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
         ) {
             unreachable!()
         }
@@ -184,7 +185,7 @@ sp_api::impl_runtime_apis! {
         fn extract_successful_bundles(
             _domain_id: DomainId,
             _extrinsics: Vec<ExtrinsicFor<Block>>,
-        ) -> sp_domains::bundle::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
+        ) -> sp_domains::bundle::OpaqueBundles<Block, DomainHeader, Balance> {
             unreachable!()
         }
