+++ crates/subspace-malicious-operator/src/malicious_bundle_tamper.rs
@@ -5,8 +5,9 @@ use sp_api::ProvideRuntimeApi;
 use sp_domain_digests::AsPredigest;
 use sp_domains::bundle::{BundleValidity, InvalidBundleType};
 use sp_domains::core_api::DomainCoreApi;
+use sp_domains::execution_receipt::BlockFees;
 use sp_domains::merkle_tree::MerkleTree;
-use sp_domains::{BlockFees, ChainId, HeaderHashingFor, OperatorPublicKey, OperatorSignature};
+use sp_domains::{ChainId, HeaderHashingFor, OperatorPublicKey, OperatorSignature};
 use sp_keystore::KeystorePtr;
 use sp_runtime::traits::{Block as BlockT, Hash as HashT, Header as HeaderT, NumberFor, One, Zero};
 use sp_runtime::{DigestItem, OpaqueExtrinsic, RuntimeAppPublic};
