+++ crates/subspace-runtime/src/lib.rs
@@ -60,10 +60,10 @@ use sp_core::crypto::KeyTypeId;
 use sp_core::{ConstBool, H256, OpaqueMetadata};
 use sp_domains::bundle::BundleVersion;
 use sp_domains::bundle_producer_election::BundleProducerElectionParams;
+use sp_domains::execution_receipt::ExecutionReceiptFor;
 use sp_domains::{
     ChannelId, DOMAIN_STORAGE_FEE_MULTIPLIER, DomainAllowlistUpdates, DomainId, DomainInstanceData,
-    ExecutionReceiptFor, INITIAL_DOMAIN_TX_RANGE, OperatorId, OperatorPublicKey,
-    PermissionedActionAllowedBy,
+    INITIAL_DOMAIN_TX_RANGE, OperatorId, OperatorPublicKey, PermissionedActionAllowedBy,
 };
 use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_domains_fraud_proof::storage_proof::{
@@ -1480,7 +1480,7 @@ impl_runtime_apis! {
         }
 
         fn submit_receipt_unsigned(
-            singleton_receipt: sp_domains::SealedSingletonReceipt<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
+            singleton_receipt: sp_domains::execution_receipt::SealedSingletonReceipt<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
         ) {
             Domains::submit_receipt_unsigned(singleton_receipt)
         }
@@ -1488,7 +1488,7 @@ impl_runtime_apis! {
         fn extract_successful_bundles(
             domain_id: DomainId,
             extrinsics: Vec<ExtrinsicFor<Block>>,
-        ) -> sp_domains::bundle::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
+        ) -> sp_domains::bundle::OpaqueBundles<Block, DomainHeader, Balance> {
             crate::domains::extract_successful_bundles(domain_id, extrinsics)
         }
