+++ domains/client/domain-operator/src/domain_block_processor.rs
@@ -20,8 +20,9 @@ use sp_core::H256;
 use sp_core::traits::CodeExecutor;
 use sp_domains::bundle::BundleValidity;
 use sp_domains::core_api::DomainCoreApi;
+use sp_domains::execution_receipt::ExecutionReceipt;
 use sp_domains::merkle_tree::MerkleTree;
-use sp_domains::{DomainId, DomainsApi, ExecutionReceipt, HeaderHashingFor};
+use sp_domains::{DomainId, DomainsApi, HeaderHashingFor};
 use sp_domains_fraud_proof::FraudProofApi;
 use sp_messenger::MessengerApi;
 use sp_mmr_primitives::MmrApi;
